Moving to Tasks instead of manually handling threading. Added Async functions
Added socket extensions to be able to use Tasks - STANDARD1_3 is a bit different because the needed socket functions are not available. Disposing of the SocketAsyncEventArgs with STANDARD1_3 in case that was creating a memory issue

Making sure that we're receiving the entire record in the ReceiveCallback function. If we are not, we buffer the received data and receive more until we have a full record

Change-Id: I932fe53b19890e17c553c3e3453d08c4bfe13397